+++ .github/CHANGELOG.md
@@ -51,6 +51,13 @@
 
 # v3.5 (Overhaul an Overhaul?)
 
+## 3.5.5
+- [**New**] Added highlight  
+  An event when a highlight word/phrase is sent by someone, the bot will send
+  you a private message with the message that triggered it along with contexts
+  if you haven't talk for 30 minutes in that channel
+  (Heavily inspired by `Highlight` by Danny/Rapptz)
+
 ## 3.5.4 (Bugfix)
 - [**Fixed**] Prefixes is not loaded properly